+++ tests/test_build_assessment_config.py
@@ -1,15 +1,11 @@
-from app.export.config_generator.generate_assessment_config import (
-    build_assessment_config,
-)
+from app.export_config.generate_assessment_config import build_assessment_config
 from tests.unit_test_data import cri1
 from tests.unit_test_data import crit_1_id
 from tests.unit_test_data import mock_form_1
 
 
 def test_build_basic_structure(mocker):
-    mocker.patch(
-        "app.export.config_generator.generate_assessment_config.get_form_for_component", return_value=mock_form_1
-    )
+    mocker.patch("app.export_config.generate_assessment_config.get_form_for_component", return_value=mock_form_1)
 
     results = build_assessment_config([cri1])
     assert "unscored_sections" in results
@@ -19,9 +15,7 @@ def test_build_basic_structure(mocker):
 
 def test_with_field_info(mocker):
 
-    mocker.patch(
-        "app.export.config_generator.generate_assessment_config.get_form_for_component", return_value=mock_form_1
-    )
+    mocker.patch("app.export_config.generate_assessment_config.get_form_for_component", return_value=mock_form_1)
     results = build_assessment_config([cri1])
     assert len(results["unscored_sections"]) == 1
     unscored_subcriteria = next(section for section in results["unscored_sections"] if section["id"] == crit_1_id)[
